conversations-service/requirements.txt
# conversation-service/requirements.txt
Flask
Flask-Injector
SQLAlchemy
psycopg2-binary
